Refactor dialog files.
